requirements.txt
requests
pyyaml
playwright
websockets
sqlalchemy
pandas
psycopg2
loguru
